Shadows fixed and video rendered

+++ libs/sdw/RayTriangleIntersection.h
@@ -8,12 +8,12 @@ struct RayTriangleIntersection {
 	glm::vec3 intersectionPoint;
 	float distanceFromCamera;
 	ModelTriangle intersectedTriangle;
-	size_t triangleIndex;
+	int triangleIndex;
 	float t;
 	float u;
 	float v;
 
 	RayTriangleIntersection();
-	RayTriangleIntersection(const glm::vec3 &point, float distance, const ModelTriangle &triangle, size_t index);
+	RayTriangleIntersection(const glm::vec3 &point, float distance, const ModelTriangle &triangle, int index);
 	friend std::ostream &operator<<(std::ostream &os, const RayTriangleIntersection &intersection);
 };
